refactor(delivery): extract producer_task + fast-delay tests to satisfy 1000-line cap
Behaviour-preserving split to get two rs-delivery files under the CI
1000-line per-file gate (aim <950):

- Move producer_task into new endpoint_producer.rs; endpoint_task.rs
  802 (was 1040). Made PrefetchedChunk, the producer-loop consts, and
  PREFETCH_BUFFER_SIZE pub(crate); call site now
  crate::endpoint_producer::producer_task.
- Move the fast-delay / chunk-gap tests into new
  fast_self_healing_tests.rs; endpoint_task_tests.rs 926 (was 1169).
  Shared mock helpers (TimedMockFetcher, MockProcessFactory,
  test_ep_cfg) made pub(crate) and reused via super::tests.

No logic or assertion changes. rs-delivery: 206 bin + 33 lib + 8
integration tests all pass, clippy clean, fmt clean.
